pyproject.toml
[project]
name = "gitlab-webhook-receiver"
version = "0.1.0"
description = "GitLab Webhook Receiver for Merge Request automation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pydantic-settings",
    "PyYAML",
    "deek-hook-review[all]",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
]
[tool.uv.sources]
deek-hook-review = { path = "/Users/anshdeep/Desktop/personal/code_reviewer_posha copy" }